Reset factorisation scales in ClusterSequenceInfo()
Fixes #358.

Otherwise, the evaluation of PDF ratios when traversing the cluster
sequence can modify the ISR handler's scale information. This in turn
will lead to the PDF info object creaed by the phase space
handler to contain one of the cluster scales instead of the hard process
scale. Note that this is entirely a book-keeping issue, given that the
PDF info object is only used for a-posteriori reweighting, and moreover
any fixed-order calculations, for which the a-posteriori reweighting
information is valid, were not affected by this.
We only fix this for consistency, and to fix #358.


(cherry picked from commit bbba1f744ee28794e6c698678742d85a618ca499)