Fix issue when using MI_HANDLER=None in FO EvtGen
Fixes #363

The behaviour of rel-2-2-x is hereby restored (there must have been a
copy&paste error swapping the check for BEAM_REMNANTS with a check for
MI_HANDLER). The difference is that we allow the user to use
BEAM_REMNANTS=Off explicitly, without then going into the alternative
branch, where the INTRINSIC_KPERP is effectively disabled.