Apply weight ratio between pilot and normal run
This is now necessary, because when using the PILOT scale setter,
the normal run will in general calculate a different weight.
The ratio to the pilot run weight must be taken into account,
either by a second unweighting, or by applying the ratio to
the resulting event weight, as we do here.