Add beam HepMC vertex even if no beam/bunch blob is present
This ensures that if the beam remnants are disabled, Rivet will still be
able to find the beams. With the logic before, there were several
particles without a production vertex, which were all implicitly seen as
beam particles by HepMC/Rivet. Now, only the two incoming (proton) beams
are thus seen, and e.g. Rivet analyses that rely on sqrtS internally
(for example, to define their binning) work as expected.