Fix Amegic for unresolved EPA photons
Previously, the boost was only set up if the ISR handler was switched on.
When we use the beam handler to create photons and these photons
are not resolved, i.e. the ISR handler is off, the boost was never
initialised and led to a divergent XS and NaNs in the momenta.
This change in the if-condition sets up the boost in the correct cases.