If a pilot-scale run is vetoed, rerun with normal scale
The reason is, that the run might not be vetoed with the normal scale,
so we need to take another round to confirm. In case there are no
variations, another rerun after unweighting will not be made, as it is
now unnecessary. However, when there are variations, we re-run again,
with all variations enabled. Hence, when using variations and a pilot
scale, there might be up to three runs until an event is fully accepted.
Note though that this should not happen too often.