Enforce consistent splitting scale reweighting
Previously, if CSS_REWEIGHT is false, but
REWEIGHT_SPLITTING_ALPHAS_SCALES or REWEIGHT_SPLITTING_PDF_SCALES are
true, the splitting scales in the CKKW history are still reweighted with
varied scales, although the shower scales are not.  This is likely not
intended by the user, and it does not make sense to vary the scales
independently.  Hence, now the two REWEIGHT_*_SCALES parameters are
forcibly set to false if CSS_REWEIGHT is false, and a warning is emitted
to notify the user, if it was explititly set.