Always consume the random number used for unweighting
This is important, because the pilot run and the re-run might consume a
different number of random numbers. But then the unweighting random
number might be reused subsequently, which however introduces a bias,
because this random number is now constrained by the unweighting
procedure.

(cherry picked from commit 8cf270121917fcc8ce61ca6311a91f8d7645f989)