Avoid numerical issues with points at large rapidities
When Beam() and ISR() is on, we sample twice in the rapidity.
Hence we can get points that lie outside the usual range of y \in [-10, 10].
This leads to numerical issues as the rapidity enters the calculation with exp(y).
This checks ensures, that the sampled rapidity lies within the usual boundaries (c.f. Collider_Kinematics.C and ISR_Handler.C) of y_max = 10.