Add "--lorentznmax" option to Sherpa-generate_model
This restricts the creation of Lorentz calculators to functions with up to <n> external particles. It is useful/necessary for the compilation of huge models (e.g. SMEFT) with extremely complicated but typically unused vertices.

If the user requests a process which requires one of the ignored vertices, it will fail with an error message. The user can then recompile the model with a higher --lorentznmax setting.