Fix for missing tau spin correlations
The fix in 3b6304a4 led to the fact where tau spins were never contracted with the tau decay matrices, not even in the actual hadron (tau) decay handler itself.

With this commit the if condition is made more specific to only cover the case where a decay handler that is not treating the tau will skip its contraction.