Introduce remnants for Pomerons
For the pomerons, we implement a strict two-particle notion, i.e. dominantly a two gluon state. In case of extracted quarks, one gluon must be created additionally to the corresponding anti-quark, to maintain the picture. In the ISR we respect light quarks, though, as included in the GKG pdf, too. MPI need to be switched off.

Updated the manual.

Let Pomeron remnants always create a gluon