Reinstate clustering in OneHEvent to fix physics issue
It would be preferable to defer this clustering to after the
unweighting, as advertised in 2209.00843, but the change introduced in
OneHEvent in 8a15a8f8a breaks physics, even though we thought it would
not, because the clustering would anyway be done later on demand as soon
as it is needed. Perhaps there is some side effect that needs to be
understood first.

(cherry picked from commit 8cb362ab627643566ebbe85c611787efc45a658a)