Polarized cross sections for massive vector bosons
This commit will enable Sherpa to simulate polarized cross sections for massive vector bosons with vector boson polarization defined in the helicity basis. The laboratory frame, the vector boson center-of-mass-frame(COM), the parton-parton-frame as well as the COM of any combination of at least two particles of the hard process can be used as reference frames for the polarization definition. The cross section of all possible polarization combinations and the interference between them are calculated in one simulation run and printed out as additional event weights.