Evaluate "None" (case-insensitive) as false for boolean settings
Before, only "false", "no", "off" and "0" were treated in this way, and
everything else was evaluated to true. This was for example relevant for
BEAM_REMNANTS, where using "None" actually enabled it. Now, it should be
disabled as expected. Hence, the user does not need to remember where to
use "false" or "None" etc.