Add PDF_VARIATION_BEAMS and PDF_VARIATION_ALPHAS_BEAM settings
PDF_VARIATION_BEAMS can be used to restrict for which beams
a PDF variation is applied. Its default is "[1, 2]", i.e. both beams
will undergo a given PDF variation. Use "1" or "2" to only apply it to a
single beam. This is a global setting for all PDF variations, i.e. it is
currently not possible to do this on the basis of a single PDF
variation.

Now that the above is added, there is an ambiguity which beams's PDF
should be used to evaluate AlphaS. For that, PDF_VARIATION_ALPHAS_BEAM
has been added.  Its default is "0", which means that the first
available beam's PDF is used. Use "1" or "2" to select a specific beam's
PDF instead.

Having different PDFs for each beam will be reflected in the variation
name (in Sherpa's output or the HepMC weight name), like in the
following example:
"MUR=1__MUF=1__LHAPDF.BEAM1=93300__LHAPDF.BEAM2=93301".