Add CMake closure test for SCALE_VARIATIONS
The test first generates events with on-the-fly scale variations and
analyses it with Rivet. Then for each variation a separate run with the
explicit variation is done. The results are then compared using `yodadiff`.
If the MC_XS/XS results are not exactly equal for all variations (within
the default tolerance of yodadiff), then the test fails. For now, this
only tests simple LO variations. More fixed-order tests will follow,
e.g. for an NLO setup. We might also add matched/merged setups with
shower emissions disabled (to allow for an exact comparison). Truely
showered variation tests can only be done statistically and will need a
corresponding setup (and perhaps way more run time to be sensible).