Fix python extension runtime error for Python 3
The previous version used PyString_Check and PyString_AsString to
convert arguments stored in char** argv. This does not work any longer
in (but compiles for) Python 3, where string are formed of characters
instead of bytes, and hence some conversion must happen. We use
PyUnicode_AsUTF8AndSize for this now.