Replace uses of sprintf -> snprintf
The reason is that sprintf is deprecated in my current toolchain
(clang-1400.0.29.202), such that I was flooded with warnings, in
particular when building AMEGIC++.

The reason for deprecation is explained here:
https://rules.sonarsource.com/c/RSPEC-6069

Note that this article states a slight performance overhead when using
snprintf instead of sprintf. However, the uses of sprintf are all about
creating formatted string, and therefore I would assume that these uses
are not performance-critical.