Fill a HepMC XS for each variation/auxiliary weight
Otherwise newer versions of HepMC (3.6 or later) will issue a warning
for each event, complaining about a mismatch of the number of weights
and the number of cross sections.

Note that the cross sections for the auxiliary weights (NTrials etc.)
are not meaningful.

Note that the cross sections for the on-the-fly variations are just set
to the nominal cross section. This is not correct, and should not be
used. The end user is responsible for doing the correct normalisation.
This is different to the rel-3 series of Sherpa, where these cross
sections are always correctly filled.