Add missing MSG_LIMIT command line equivalent
It was already added to the command line enum, but there was no option
associated with it yet. I have chosen `-W`. Note that without specifying
an accompanying option, it is useless to add settings to the command
line enum. Settings given as `X=Y` or `X:Y` or `"X: Y"` etc. will always
be read from the command line regardless of whether something has been
added to Command_Line_Options.H or not. This is only to add associated
command line options a la `-x|--long-x`.