Fix bug when MPI merging split rivet analyses
Namely, when one of the ranks never saw an event for the given split
analysis, it could not properly initialise a appropriate Rivet analysis
handler with a dummy event, because the run info of the dummy event was
not deep copied. This is now remedied by explicitly copying the run
info.