Fix num. issue in AMEGIC ext. state construction
This basically resets transverse components of incoming particles to
zero to remove numerical inaccuracies. Also, for massless particles, the
energies are reset to the mass shell value. This was necessary in our
TTBar validation effort for v.3.0.0, where otherwise the initial state
construction in Amegic failed to work properly, as it does exact checks
against zero when it decides what auxiliary vector to use.