Fix fallback for -DHepMC3_DIR
The correct form is -DHepMC3_DIR, but since it is the only
such package name that is not completely capitalised, errors
are bound to happen on the user side. Therefore, commit
1389d274 included a fallback in the build system so
that the user could also use the fully uppercase version.
The fallback turned out not to work, which is fixed in
this commit.