Silence harmless MPI static compilation warning
A warning was emitted, because the sendbuf argument passed to
MPI_Allgatherv had type (const char*), but this is inconsistent with the
MPI data type MPI_CHAR given as the send data type. However, the sendbuf
argument is actually of type (const void*), so it is safe to cast (const
char*) to (char*) here, to silence the warning.