Check first whether linker flag -single_module is obsolete
... before using it to compile/link shared libraries on a Darwin system
(i.e. MacOS). This linker flag has become obsolete recently, and since
then warnings about that have been emitted during compilation. Now, we
check first if the flag is obsolete, and only then add it to list of
flags used. To do that, we add and use the CMake macro
CheckLinkerFlagRegex.