Revert project name of the manual to "Sherpa Manual"
It was changed previously from "Sherpa Manual" to "Manual" when the logo
was changed (the new version of the logo already contains the string
"SHERPA", and the default sphinx theme for HTML output we use displays
the project name immediately below the logo; hence, if the project name
includes the word "Sherpa", this leads to an immediate repetition).

However, this change broke a link to the PDF version in the manual
deployed to the project's GitLab pages. Also, the PDF version of the
manual was then simply titled "Manual", which is not good.

Hence, we now revert this change, using the proper project name "Sherpa
Manual" again. The problem with the immediate repetition of the word
"Sherpa" in the HTML version is fixed by setting the HTML theme option
"logo_name" to False, which removes the project name display below the
logo entirely.


(cherry picked from commit 461a5aa0056a11cb9483cdff3108caf0a9c865f5)

Co-authored-by: Enrico Bothmann <enrico.bothmann@uni-goettingen.de>