Fix mistake in combination of variations
Fixes #617.

Before, when combining parton-shower and "main" (= matrix-element)
variations, the resulting full variation was not completely correct
(although close to the correct value), whenever weights map were added
together (which is always the case when generating more than one event,
or perhaps one event on more than one MPI rank), because the order of
this addition and the multiplication for the combination was wrong.

Now, we avoid the problematic combination by just keeping track of the
full (parton-shower & main) variation separately.

Note that the issue might not have been relevant for most practical
purposes, because e.g. Rivet uses per-event information, which was
correct. Only the internal book-keeping was slightly off, which is for
example used for the cross section table at the end of a run.

(cherry picked from commit e94bd66c05ceed1550e195029cd367d78471382b)