Revert activating diquarks [physval]
They were disabled first in aa2e81ad85
when the line `Hadron_Init().Init()`
(which adds diquarks, but disables them).
was added early in InitializeTheFramework(),
because this then took precedence over
the diquark initialization from the beams
(which adds the diquarks, and enables them).
I do not know if this change was deliberate,
but since #632 is not about enabling/disabling
diquarks, I revert enabling them to again
duplicate the behaviour on master.