Revert to majorana=-1 for self-anti hadrons [physval]
I.e. set the majorana property to -1 for those hadrons, that are their
own antiparticle. I am not sure why Flavour::IsAnti is defined as it is
(seems confusing and dangerous to use m_majorana to encode both if a
particle is a Majorana fermion, but also if a particle is a hadron and
boson which is its own anti-particle, by using the magical value `-1`),
but fixing this is not in the scope of this MR.