Guard against PRINT_MPI_XS=1 & Event_Files=HDF5[]
Guard against the setting that will, in general, cause synchronization
to break in MPI runs, such that the entire run stalls.

The reason is that currently the synchronisation between the HDF5
readers and Event_Handler via the common MPI_Object base class and its
Communicate function assumes that each ranks calls Communicate from
HDF5_Reader a varying number of times, followed by a varying number of
Communicate calls from Event_Handler at the end of the run. With
PRINT_MPI_XS=1, this assumption is not correct any longer.