Add newlines if output interrupts progress output
Progress output means that one or more lines are to be updated or
appended to, like the event generation progress updates, or the
setting-up-processes progress with the dots being printed one by one.

Previously, any other output would interrupt this in a non-graceful way,
e.g. after a carriage return it would just override the current progress
line, or it would come in between the dots of the setting-up-progresses
output, etc. This is now prevented by first adding n newlines before
printing the interrupting output.