Omit pilot run warning if both weights are NaN
Sherpa prints a warning if the nominal result from the pilot run and the
nominal result from the rerun are not equal. A false positive warning
was printed when both results were NaN, due to how NaN comparisons work.
Now, this case is explicitly checked, and a warning is not printed.